[AI/fix] 인증관련 페이지 함수 화면 challenge/certification 디렉터리로 옮김
