+++ frontend/lib/challenge/certification/run_model_by_image_demo.dart
