+++ frontend/lib/challenge/certification/ui/box_widget.dart
