+++ frontend/lib/challenge/certification/ui/camera_view.dart
@@ -1,10 +1,10 @@
 import 'dart:developer';
 import 'dart:io';
-
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:pytorch_lite/pytorch_lite.dart';
+import 'package:image/image.dart' as img;
 
 import 'camera_view_singleton.dart';
