+++ frontend/lib/challenge/certification/ui/camera_view_singleton.dart
