+++ frontend/lib/main.dart
@@ -1,71 +1,7 @@
 import 'package:flutter/material.dart';
-import 'run_model_by_camera_demo.dart';
-import 'run_model_by_image_demo.dart';
+import 'package:frontend/challenge/certification/certifacation_screen.dart';
+
 
 Future<void> main() async {
-  runApp(const ChooseDemo());
-}
-
-class ChooseDemo extends StatefulWidget {
-  const ChooseDemo({Key? key}) : super(key: key);
-
-  @override
-  State<ChooseDemo> createState() => _ChooseDemoState();
-}
-
-class _ChooseDemoState extends State<ChooseDemo> {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('정해진 자세를 취해주세요!'),
-        ),
-        body: Builder(builder: (context) {
-          return Center(
-            child: Column(
-              children: [
-                TextButton(
-                  onPressed: () => {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                          builder: (context) => const RunModelByCameraDemo()),
-                    )
-                  },
-                  style: TextButton.styleFrom(
-                    backgroundColor: Colors.blue,
-                  ),
-                  child: const Text(
-                    "카메라",
-                    style: TextStyle(
-                      color: Colors.white,
-                    ),
-                  ),
-                ),
-                TextButton(
-                  onPressed: () => {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                          builder: (context) => const RunModelByImageDemo()),
-                    )
-                  },
-                  style: TextButton.styleFrom(
-                    backgroundColor: Colors.blue,
-                  ),
-                  child: const Text(
-                    "갤러리",
-                    style: TextStyle(
-                      color: Colors.white,
-                    ),
-                  ),
-                )
-              ],
-            ),
-          );
-        }),
-      ),
-    );
-  }
+  runApp(const CertifacationScreen());
 }
